store/serializers.py
from rest_framework import serializers
from .models import Address, Coordinates

class AddressSerializer(serializers.ModelSerializer):
  class Meta:
    model = Address
    fields = '__all__'

class CoordinatesSerializer(serializers.ModelSerializer):
  class Meta:
    model = Coordinates
    fields = '__all__'