requirements.txt
aiohttp==3.6.2
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Django==3.0
django-cors-headers==3.2.0
djangorestframework==3.11.0
haversine==2.1.2
idna==2.8
lxml==4.4.2
multidict==4.7.1
pygeocoder==1.2.5
pytz==2019.3
requests==2.22.0
simplejson==3.17.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
yarl==1.4.2
psycopg2==2.6.2
